requirements.txt
# Core dependencies
boto3>=1.26.0
tqdm>=4.65.0
pydantic>=2.0.0
openai>=1.0.0
nltk>=3.8.0
pandas>=2.0.0
pyarrow>=14.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0

# Optional but recommended
fastparquet>=2023.10.1
wordcloud>=1.9.0
